requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
huey==2.4.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
Markdown==3.4.3
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
prompt-toolkit==3.0.39
psycopg==3.1.9
psycopg-binary==3.1.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
